% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcbc.R
\name{bcrange}
\alias{bcrange}
\title{Range scaling batch correction}
\usage{
bcrange(peak, qc = NULL, log = T)
}
\arguments{
\item{peak}{peaks intensity across samples}

\item{qc}{peaks intensity pooled QC samples}

\item{log}{log transformation}
}
\value{
corrected peaks intensity
}
\description{
Range scaling batch correction
}
