% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcbc.R
\name{bcscaling}
\alias{bcscaling}
\title{Scaling batch correction}
\usage{
bcscaling(peak, qc = NULL, log = TRUE)
}
\arguments{
\item{peak}{peaks intensity across samples}

\item{qc}{peaks intensity pooled QC samples}

\item{log}{log transformation}
}
\value{
corrected peaks intensity
}
\description{
Scaling batch correction
}
