% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{isvacor}
\alias{isvacor}
\title{Use Independent Surrogate Variable Analysis(ISVA) to correct the unknown batch effects}
\usage{
isvacor(data, lv)
}
\arguments{
\item{data}{data as mzrt profile}

\item{lv}{factor vector for the group infomation}
}
\value{
list object with raw data, corrected data, signal part, batch part, random errors part,  p-values, q-values. If no independent surrogate variable found, corresponding part would miss.
}
\description{
Use Independent Surrogate Variable Analysis(ISVA) to correct the unknown batch effects
}
\details{
Use Independent Surrogate Variable Analysis(ISVA) to correct the unknown batch effects
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
list <- getmr(cdfpath, pmethod = ' ')
list <- isvacor(list$data,list$group$class)
}
}
\seealso{
\code{\link{svacor}}, \code{\link{pcacor}},\code{\link{limmafit}},\code{\link{famtcor}}
}
