% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{limmafit}
\alias{limmafit}
\title{LIMMA to test batch effects by p-value and q-value}
\usage{
limmafit(data, lv, batch = NULL)
}
\arguments{
\item{data}{data as mzrt profile}

\item{lv}{vector for the group information}

\item{batch}{vector for the batch information}
}
\value{
list object with raw data, corrected data, signal part, batch part, random errors part,  p-values, q-values. If no batch index, corresponding part would miss.
}
\description{
LIMMA to test batch effects by p-value and q-value
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
list <- getmr(cdfpath, pmethod = ' ')
li <- limmafit(list$data,list$group$class)
}
}
\seealso{
\code{\link{isvacor}},\code{\link{svacor}}, \code{\link{pcacor}},\code{\link{famtcor}}
}
