% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{limmaplot}
\alias{limmaplot}
\title{Filter the data with p value and q value and show them as heatmap}
\usage{
limmaplot(list, lv, pt = 0.05, qt = 0.05, index = NULL)
}
\arguments{
\item{list}{results from `limmafit` function or related batch correction method}

\item{lv}{factor vector for the group infomation}

\item{pt}{threshold for p value, default is 0.05}

\item{qt}{threshold for q value, default is 0.05}

\item{index}{index for selected peaks}
}
\value{
heatmap for the data
}
\description{
Filter the data with p value and q value and show them as heatmap
}
\examples{
\dontrun{
sim <- mzrtsim()
li <- svacor(log(sim$data), as.factor(sim$con))
limmaplot(li,as.factor(sim$con))
}
}
\seealso{
\code{\link{isvacor}},\code{\link{svacor}}, \code{\link{pcacor}},\code{\link{limmafit}}
}
