% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{limmaroc}
\alias{limmaroc}
\title{Plot ROC curve for simulation data}
\usage{
limmaroc(sim, pvalue, points = 100)
}
\arguments{
\item{sim}{simulation results from `mzrtsim` or `simmzrt`}

\item{pvalue}{p value results from `limmafit` function or related batch correction method}

\item{points}{numbers of points for ROC curve}
}
\description{
Plot ROC curve for simulation data
}
