% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrtsim.R
\docType{data}
\name{monahrms1}
\alias{monahrms1}
\title{A list containing MoNA LC-HRMS spectra database}
\format{
A list with compounds from MoNA LC-HRMS spectra database for LC-HRMS simulation
}
\usage{
data(monahrms1)
}
\description{
A list containing MoNA LC-HRMS spectra database
}
\keyword{datasets}
