% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzmlsim.R
\docType{data}
\name{mzm}
\alias{mzm}
\title{A vector containing m/z from matrix blank}
\format{
A vector containing m/z from matrix blank for LC-HRMS simulation
}
\usage{
data(mzm)
}
\description{
A vector containing m/z from matrix blank
}
\keyword{datasets}
