% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzmlsim.R
\name{mzmlviz}
\alias{mzmlviz}
\title{Generate figure for mzml raw data}
\usage{
mzmlviz(mzml, mzrange = c(100, 1000), rtrange = c(0, 600))
}
\arguments{
\item{mzml}{mzML file path}

\item{mzrange}{m/z range for simulation, peaks out of the range will be removed, default c(100,1000)}

\item{rtrange}{retention time range for simulation, default c(0,600)}
}
\value{
figure for raw data and dataframe with mz, rt, and intensity
}
\description{
Generate figure for mzml raw data
}
