% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrtsim.R
\name{mzrtsim}
\alias{mzrtsim}
\title{Generate simulated count data with batch effects for npeaks}
\usage{
mzrtsim(
  ncomp = 100,
  fc = NULL,
  ncond = 2,
  ncpeaks = 0.1,
  nbatch = 3,
  nbpeaks = 0.1,
  npercond = 10,
  nperbatch = c(8, 5, 7),
  batchtype = "mb",
  rtsim = TRUE,
  db = NULL,
  seed = 42
)
}
\arguments{
\item{ncomp}{compound numbers to be selected, default 100}

\item{fc}{fold change of compounds with the same length of compounds numbers, default NULL}

\item{ncond}{Number of conditions to simulate}

\item{ncpeaks}{percentage of compounds influenced by conditions}

\item{nbatch}{Number of batches to simulate}

\item{nbpeaks}{percentage of peaks influenced by batchs}

\item{npercond}{Number of samples per condition to simulate}

\item{nperbatch}{Number of samples per batch to simulate}

\item{batchtype}{type of batch. 'm' means monotonic, 'b' means block, 'r' means random error, 'mb' means mixed mode of monotonic and block, default 'mb'}

\item{rtsim}{logical, simulate retention time or not}

\item{db}{compound database with MS1 data. e.g. hmdbcms or monams1}

\item{seed}{Random seed for reproducibility}
}
\value{
list with rtmz data matrix, row index of peaks influenced by conditions, row index of peaks influenced by batchs, column index of conditions, column of batchs, raw condition matrix, raw batch matrix, peak mean across the samples, peak rsd across the samples
}
\description{
Generate simulated count data with batch effects for npeaks
}
\details{
the numbers of batch columns should be the same with the condition columns.
}
\examples{
data(hmdbcms)
sim <- mzrtsim(db=hmdbcms)
}
\seealso{
\code{\link{simdata}}
}
