% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{pcacor}
\alias{pcacor}
\title{Use Principal component analysis(PCA) to correct the unknown batch effects}
\usage{
pcacor(data, lv)
}
\arguments{
\item{data}{data as mzrt profile}

\item{lv}{factor vector for the group infomation}
}
\value{
list object with various components such raw data, corrected data, signal part, random errors part, batch part, p-values, q-values.
}
\description{
Use Principal component analysis(PCA) to correct the unknown batch effects
}
\details{
Use Principal component analysis(PCA) to correct the unknown batch effects
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
list <- getmr(cdfpath, pmethod = ' ')
li <- pcacor(list$data,list$group$class)
}
}
\seealso{
\code{\link{isvacor}},\code{\link{svacor}}, \code{\link{limmafit}},\code{\link{famtcor}}
}
