% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{rrmixcor}
\alias{rrmixcor}
\title{Use Random main effect and Random compound-specific error variance with a mixture structure(RRmix) to correct the unknown batch effects}
\usage{
rrmixcor(data, lv)
}
\arguments{
\item{data}{data as mzrt profile}

\item{lv}{factor vector for the group infomation}
}
\value{
list object with various components such raw data, posterior probability.
}
\description{
Use Random main effect and Random compound-specific error variance with a mixture structure(RRmix) to correct the unknown batch effects
}
\details{
Use Random main effect and Random compound-specific error variance with a mixture structure(RRmix) to correct the unknown batch effects
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
list <- getmr(cdfpath, pmethod = ' ')
li <- rrmixcor(list$data,list$group$class)
}
}
\seealso{
\code{\link{isvacor}},\code{\link{svacor}}, \code{\link{limmafit}},\code{\link{pcacor}}
}
