% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzmlsim.R
\name{simmzml_blank}
\alias{simmzml_blank}
\title{Generate simulated mzml data for blank sample}
\usage{
simmzml_blank(
  name,
  mzrange = c(100, 1000),
  rtrange = c(0, 600),
  scanrate = 0.2,
  baseline = 100,
  baselinesd = 30,
  mzdigit = 5,
  matrixmz = NULL
)
}
\arguments{
\item{name}{file name of mzml}

\item{mzrange}{m/z range for simulation, peaks out of the range will be removed, default c(100,1000)}

\item{rtrange}{retention time range for simulation, default c(0,600)}

\item{scanrate}{time for each full scan, default 0.2 second or 5 spectra per second}

\item{baseline}{noise baseline, default 100. If it's a numeric vector, it will be used as the baseline for each scan.}

\item{baselinesd}{standard deviation for noise, default 30}

\item{mzdigit}{m/z digits, default 5}

\item{matrixmz}{custom matrix m/z vector, default NULL and predefined list from serum blank will be used.}
}
\value{
one mzML file for simulated blank data
}
\description{
Generate simulated mzml data for blank sample
}
\examples{
\donttest{
simmzml_blank(name = 'blank_test')
}
}
