% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrtsim.R
\name{simplot}
\alias{simplot}
\title{Plot the heatmap for the data}
\usage{
simplot(data, lv, index = NULL)
}
\arguments{
\item{data}{matrix for heatmap}

\item{lv}{factor vector for the group infomation}

\item{index}{row index for the data matrix to be ploted}
}
\description{
Plot the heatmap for the data
}
