% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrtsim.R
\name{simroc}
\alias{simroc}
\title{Plot ROC curve for simulation data}
\usage{
simroc(sim, points = 100)
}
\arguments{
\item{sim}{list from `mzrtsim` or `simmzrt`}

\item{points}{numbers of points for ROC curve}
}
\description{
Plot ROC curve for simulation data
}
