% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{svacor}
\alias{svacor}
\title{Use Surrogate Variable Analysis(SVA) to correct the unknown batch effects}
\usage{
svacor(data, lv)
}
\arguments{
\item{data}{data as mzrt profile}

\item{lv}{factor vector for the group infomation}
}
\value{
list object with raw data, corrected data, signal part, batch part, random errors part,  p-values, q-values. If no surrogate variables were found, corresponding part would miss.
}
\description{
Use Surrogate Variable Analysis(SVA) to correct the unknown batch effects
}
\details{
Use Surrogate Variable Analysis(SVA) to correct the unknown batch effects
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
list <- getmr(cdfpath, pmethod = ' ')
li <- svacor(list$data,list$group$class)
}
}
\seealso{
\code{\link{isvacor}}, \code{\link{pcacor}},\code{\link{limmafit}},\code{\link{famtcor}}
}
