% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzmlsim.R
\name{simmzml}
\alias{simmzml}
\title{Generate simulated mzml data and compounds list}
\usage{
simmzml(
  db,
  name,
  n = 100,
  inscutoff = 0.05,
  mzrange = c(100, 1000),
  rtrange = c(0, 600),
  ppm = 5,
  mzdigit = 5,
  scanrate = 0.2,
  pwidth = 10,
  baseline = 100,
  baselinesd = 30,
  sn = 100,
  tailingfactor = 1.2,
  compound = NULL,
  rtime = NULL,
  tailingindex = NULL,
  seed = 42,
  unique = FALSE,
  matrix = FALSE,
  matrixmz = NULL
)
}
\arguments{
\item{db}{compound database with MS1 data. e.g. hmdbcms or monams1}

\item{name}{file name of mzml}

\item{n}{compound numbers from database, if compound is not NULL, n will be compound number, default 100}

\item{inscutoff}{intensity cutoff for MS1 spectra, default 0.05}

\item{mzrange}{m/z range for simulation, peaks out of the range will be removed, default c(100,1000)}

\item{rtrange}{retention time range for simulation, default c(0,600)}

\item{ppm}{m/z shift in ppm, default 5}

\item{mzdigit}{m/z digits, default 5}

\item{scanrate}{time for each full scan, default 0.2 second or 5 spectra per secound}

\item{pwidth}{peak width for the compound. If it's a single value, simulated peaks' width will use this number as the lambda of Poisson distribution. If it's a numeric vector, it will be used as the peak width for each compounds.}

\item{baseline}{noise baseline, default 100}

\item{baselinesd}{standard deviation for noise, default 30}

\item{sn}{signal to noise ratio of each compound, default 100 for all compounds when baseline is 100}

\item{tailingfactor}{tailing factor for peaks, larger means larger tailing, default 1.2}

\item{compound}{numeric compound index in the database for targeted analysis, default NULL}

\item{rtime}{retention time for the compounds, if NULL, retention time will be simulated. Default NULL}

\item{tailingindex}{numeric index for tailing compounds, if NULL, all peaks will tailing. Default NULL}

\item{seed}{Random seed for reproducibility}

\item{unique}{if TRUE, one compound will have one spectra. Default FALSE}

\item{matrix}{if TRUE, m/z from experimental data will be used for background m/z simulation.Default FALSE}

\item{matrixmz}{custom matrix m/z vector, default NULL and predefined list will be used.}
}
\value{
one mzML file for simulated data and one csv file the simulated compounds with retention time, m/z and name
}
\description{
Generate simulated mzml data and compounds list
}
\examples{
data(monams1)
simmzml(db=monams1, name = 'test')
}
